Allow services to use the api renderer
By converting dict or list types to a Bundle in the renderer,
services can now also use the api renderer.

For now, the same templates obj_list and obj_detail will be used,
but that is likely to change later as resources will contain a lot of
specific stuff that doesn't apply to services they may require
different templates.