Deal with flash errors from the login form.
In order for errors to show, we still need to add the error class to the form
if there are validation errors, in order to do this we use {% set %} in the
template to create a temporary variable.

Also, on failed login, just create a new form, so the previous username
doesn't get pre-filled.