Refactor for services, schema is now defined on the @view decorator
(instead of the Service meta class as it was before)

Had a look how Cornice does this first, however Cornice does not deal with multiple @view decorators, each with different schemas on the same view function, this is currently a design flaw in cornice.

Made use of a new Pyramid 1.7 feature "custom view derivers" in order to achieve this in a pretty clean way which is cool, and using multiple @view decorators with different schemas on the same view is allowed.

Also renamed vindaloo.decorator to vindaloo.decorators

Also fixed a bug in service, was not calling .lower() on verb.